# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/integrations/1c/sync/order/{orderNumber}": {
      "post": {
        "operationId": "IntegrationController_syncOrder",
        "parameters": [
          {
            "name": "orderNumber",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Забрать один заказ из 1С по номеру",
        "tags": [
          "Integration - 1C"
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List orders",
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create order",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get order by ID",
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update order (field-level RBAC)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/status": {
      "post": {
        "operationId": "OrdersController_transitionStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transition order status (state machine + RBAC + audit)",
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Transition order status (alias for POST :id/status)",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/production-stages/{code}": {
      "patch": {
        "operationId": "OrdersController_updateProductionStage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update production stage status (Канбан цеха)",
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "ЦМК АВРОРА — ERP REST API",
    "description": "REST API v1 ЦМК АВРОРА — планирование производства, заказы, склад, финансы",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}